FEAT: S3

+++ src/main/java/someone/alcoholic/service/heart/HeartServiceImpl.java
@@ -12,9 +12,8 @@ import someone.alcoholic.exception.CustomRuntimeException;
 import someone.alcoholic.repository.heart.HeartRepository;
 import someone.alcoholic.service.board.BoardService;
 import someone.alcoholic.service.member.MemberService;
-import someone.alcoholic.service.token.TokenService;
 
-import javax.servlet.http.HttpServletRequest;
+import java.security.Principal;
 
 @Slf4j
 @Service
@@ -24,35 +23,33 @@ public class HeartServiceImpl implements HeartService {
     private final HeartRepository heartRepository;
     private final BoardService boardService;
     private final MemberService memberService;
-    private final TokenService tokenService;
 
-    public void saveBoardHeart(HttpServletRequest request, Long boardSeq) {
-        log.info("{} 게시글 좋아요 등록 요청", boardSeq);
-        Member member = getMemberFromToken(request);
+    public void saveBoardHeart(Principal principal, Long boardSeq) {
         Board board = boardService.findBoardBySeq(boardSeq);
+        Member member = board.getMember();
+        String memberId = member.getId();
+        log.info("{} 유저 - {} 게시글 좋아요 등록 요청", memberId, boardSeq);
+        memberService.checkAuthorizedUser(principal, member.getId());
         if (heartRepository.existsByMemberAndBoard(member, board)) {
-            log.info("{} 게시글 좋아요 등록 실패 - 이미 좋아요 등록된 유저", boardSeq);
+            log.info("{} 유저 -{} 게시글 좋아요 등록 실패 - 이미 좋아요 등록된 유저", memberId, boardSeq);
             throw new CustomRuntimeException(HttpStatus.BAD_REQUEST, ExceptionEnum.HEART_ALREADY_EXISTS);
         }
         Heart heart = Heart.builder().member(member).board(board).build();
         heartRepository.save(heart);
-        log.info("{} 유저 {} 게시글 좋아요 등록 성공", member.getId(), boardSeq);
+        log.info("{} 유저 - {} 게시글 좋아요 등록 성공", memberId, boardSeq);
     }
 
-    private Member getMemberFromToken(HttpServletRequest request) {
-        String memberId = tokenService.getMemberIdByAccessToken(request);
-        return memberService.findMemberById(memberId);
-    }
-
-    public void deleteBoardHeart(HttpServletRequest request, Long boardSeq) {
-        log.info("{} 게시글 좋아요 삭제 요청", boardSeq);
-        Member member = getMemberFromToken(request);
+    public void deleteBoardHeart(Principal principal, Long boardSeq) {
         Board board = boardService.findBoardBySeq(boardSeq);
+        Member member = board.getMember();
+        String memberId = member.getId();
+        log.info("{} 유저 - {} 게시글 좋아요 삭제 요청", memberId, boardSeq);
+        memberService.checkAuthorizedUser(principal, member.getId());
         if (!heartRepository.existsByMemberAndBoard(member, board)) {
-            log.info("{} 게시글 좋아요 삭제 실패 - 좋아요 등록하지 않은 유저", boardSeq);
+            log.info("{} 유저 -{} 게시글 좋아요 삭제 실패 - 좋아요 등록하지 않은 유저", memberId, boardSeq);
             throw new CustomRuntimeException(HttpStatus.BAD_REQUEST, ExceptionEnum.HEART_NOT_EXISTS);
         }
         heartRepository.deleteById(boardSeq);
-        log.info("{} 유저 {} 게시글 좋아요 삭제 성공", member.getId(), boardSeq);
+        log.info("{} 유저 - {} 게시글 좋아요 삭제 성공", memberId, boardSeq);
     }
 }
