+++ src/main/java/someone/alcoholic/service/member/MemberServiceImpl.java
@@ -11,10 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import someone.alcoholic.domain.member.Member;
 import someone.alcoholic.domain.token.RefreshToken;
-import someone.alcoholic.dto.member.AccountDto;
-import someone.alcoholic.dto.member.MemberDto;
-import someone.alcoholic.dto.member.MemberSignupDto;
-import someone.alcoholic.dto.member.NicknameDto;
+import someone.alcoholic.dto.member.*;
 import someone.alcoholic.enums.*;
 import someone.alcoholic.exception.CustomRuntimeException;
 import someone.alcoholic.repository.NicknameRepository;
@@ -43,9 +40,7 @@ public class MemberServiceImpl implements MemberService {
 
     @Value("${image.member.directory}")
     private String imageDirectory;
-    @Value("${image.member.profile.directory}")
-    private String profileDirectory;
-    @Value("${image.member.profile.default}")
+    @Value("${image.member.default}")
     private String profileDefaultImage;
     private final MemberRepository memberRepository;
     private final TmpMemberRepository tmpMemberRepository;
@@ -59,17 +54,16 @@ public class MemberServiceImpl implements MemberService {
     @Transactional
     @Override
     public Member signup(MemberSignupDto signupDto) {
-        log.info("member sign up 시작");
+        log.info("회원가입 시작");
         checkDuplicatedId(signupDto);
         mailService.checkEmailCertified(MailType.SIGNUP, signupDto.getEmail());
-
         String randomAdj = nicknameRepository.findRandomAdj();
         String randomNoun = nicknameRepository.findRandomNoun();
         int randomNum = ThreadLocalRandom.current().nextInt(99999);
         String nickname = randomNum + "번 " + randomAdj + "한 " + randomNoun;
         return memberRepository.save(Member.createLocalMember(
                 signupDto.getId(), passwordEncoder.encode(signupDto.getPassword()),
-                nickname, FileUtil.buildDefaultFilePath(imageDirectory, profileDirectory, profileDefaultImage), signupDto.getEmail()));
+                nickname, FileUtil.buildDefaultFilePath(imageDirectory, profileDefaultImage), signupDto.getEmail()));
     }
 
     @Transactional
@@ -83,11 +77,10 @@ public class MemberServiceImpl implements MemberService {
         if(tokenClaims == null) {
             throw new CustomRuntimeException(HttpStatus.BAD_REQUEST, ExceptionEnum.TOKEN_NOT_EXIST);
         }
-
         String memberId = tokenClaims.get(AuthToken.MEMBER_ID, String.class);
         Member member = tmpMemberRepository.findById(memberId)
                 .orElseThrow(() -> new CustomRuntimeException(HttpStatus.BAD_REQUEST, ExceptionEnum.TMP_USER_NOT_EXIST))
-                .convertToMember(nickname, FileUtil.buildDefaultFilePath(imageDirectory, profileDirectory, profileDefaultImage));
+                .convertToMember(nickname, FileUtil.buildDefaultFilePath(imageDirectory, profileDefaultImage));
         memberRepository.save(member);
 
         UUID refreshTokenPk = UUID.randomUUID();
@@ -129,78 +122,75 @@ public class MemberServiceImpl implements MemberService {
 
     // 비밀번호 변경
     @Transactional(rollbackFor = {Exception.class})
-    public void changeMemberPassword(Principal principal, AccountDto accountDto) {
-        String loginId = principal.getName();
-        log.info("{} 멤버 - 비밀번호 변경 요청 시작", loginId);
-        String id = accountDto.getId();
-        identifyLoginUser(loginId, id);
+    public void changeMemberPassword(Principal principal, String id, newPasswordChangeDto passwordChangeDto) {
+        log.info("{} 멤버 - 비밀번호 변경 요청 시작", id);
+        checkAuthorizedUser(principal, id);
         Member member = findMemberById(id);
         String password = member.getPassword();
-        String inputPassword = accountDto.getPassword();
+        String inputPassword = passwordChangeDto.getPassword();
         checkCurrentPasswordEqual(id, password, inputPassword);
-        changePassword(member, accountDto.getNewPassword());
+        changePassword(member, passwordChangeDto.getNewPassword());
         log.info("{} 비밀번호 변경 요청 완료", id);
     }
 
-
-
     // 비밀번호 초기화
     @Transactional(rollbackFor = {Exception.class})
-    public void resetMemberPassword(AccountDto accountDto) {
-        String id = accountDto.getId();
-        String email = accountDto.getEmail();
-        log.info("비밀번호 초기화 요청 시작 - id : {}, email : {}", id, email);
+    public void resetMemberPassword(String id, newPasswordResetDto passwordResetDto) {
+        log.info("비밀번호 초기화 요청 시작 - id : {}", id);
+        String email = passwordResetDto.getEmail();
+        log.info("비밀번호 초기화 요청 시작 - email : {}", email);
         mailService.checkEmailCertified(MailType.PASSWORD, email);
         Member member = getMemberByIdAndEmail(id, email);
-        changePassword(member, accountDto.getNewPassword());
+        changePassword(member, passwordResetDto.getNewPassword());
         log.info("비밀번호 초기화 완료 - id : {}, email : {}", id, email);
     }
 
     @Transactional(rollbackFor = {Exception.class})
-    public void changeMemberNickname(Principal principal, String memberId, NicknameDto nicknameDto) {
-        String loginId = principal.getName();
-        log.info("{} 유저 닉네임 변경 요청 시작", loginId);
-        identifyLoginUser(loginId, memberId);
+    public void changeMemberNickname(Principal principal, String id, NicknameDto nicknameDto) {
+        log.info("{} 유저 닉네임 변경 요청 시작", id);
+        checkAuthorizedUser(principal, id);
         String newNickname = nicknameDto.getNickname();
-        Member member = findMemberById(memberId);
+        Member member = findMemberById(id);
         String currentNickName = member.getNickname();
-        log.info("{} 유저 닉네임 - current : {} -> new : {}", memberId, currentNickName, newNickname);
+        log.info("{} 유저 닉네임 - current : {} -> new : {}", id, currentNickName, newNickname);
         checkDuplicateNickname(newNickname);
         member.setNickname(newNickname);
         Member savedMember = memberRepository.save(member);
-        log.info("{} 유저 닉네임 변경 요청 완료 - nickName : {} -> {}", memberId, currentNickName, savedMember.getNickname());
+        log.info("{} 유저 닉네임 변경 요청 완료 - nickName : {} -> {}", id, currentNickName, savedMember.getNickname());
     }
 
     @Transactional(rollbackFor = {Exception.class})
-    public void changeMemberImage(Principal principal, String memberId, MultipartFile multipartFile) {
-        String loginId = principal.getName();
-        log.info("{} 유저 이미지 변경 요청 시작 - file : {}", loginId, multipartFile.getName());
-        identifyLoginUser(loginId, memberId);
-        Member member = findMemberById(memberId);
+    public void changeMemberImage(Principal principal, String id, MultipartFile multipartFile) {
+        log.info("{} 유저 이미지 변경 요청 시작 - file : {}", id, multipartFile.getName());
+        checkAuthorizedUser(principal, id);
+        Member member = findMemberById(id);
         String memberImageUrl = member.getImage();
-        log.info("{} 유저 currentImage : {}", memberId, memberImageUrl);
-        String savedImagePath = s3Uploader.uploadFile(multipartFile, member.getSeq(), memberImageUrl, imageDirectory, profileDirectory);
-        member.setImage(savedImagePath);
+        log.info("{} 유저 currentImage : {}", id, memberImageUrl);
+        FileUtil.validateFile(multipartFile); // 파일 존재 체크
+        s3Uploader.deleteFile(memberImageUrl); // 기존 파일 삭제
+        String filePath = FileUtil.buildFilePath(multipartFile, imageDirectory, member.getSeq()); //저장 경로
+        s3Uploader.uploadFile(multipartFile, filePath);
+        member.setImage(filePath);
         memberRepository.save(member);
-        log.info("{} 유저 이미지 변경 완료 - path : {}", memberId, savedImagePath);
+        log.info("{} 유저 이미지 변경 완료 - path : {}", id, filePath);
     }
 
     @Transactional(rollbackFor = {Exception.class})
-    public void deleteMemberImage(Principal principal, String memberId) {
-        String loginId = principal.getName();
-        log.info("{} 유저 이미지 삭제 요청 시작", loginId);
-        Member member = findMemberById(memberId);
+    public void deleteMemberImage(Principal principal, String id) {
+        log.info("{} 유저 이미지 삭제 요청 시작", id);
+        checkAuthorizedUser(principal, id);
+        Member member = findMemberById(id);
         String memberImageUrl = member.getImage();
-        log.info("{} 유저 currentImage : {}", memberId, memberImageUrl);
+        log.info("{} 유저 currentImage : {}", id, memberImageUrl);
         if(!s3Uploader.deleteFile(memberImageUrl)) {
             log.info("{} 유저 기본 이미지 상태에서 삭제 요청");
             throw new CustomRuntimeException(HttpStatus.BAD_REQUEST, ExceptionEnum.FILE_REMOVE_NOT_ALLOWED);
         }
         // 기본 이미지로 변경
-        String defaultImagePath = FileUtil.buildDefaultFilePath(imageDirectory, profileDirectory, profileDefaultImage);
+        String defaultImagePath = FileUtil.buildDefaultFilePath(imageDirectory, profileDefaultImage);
         member.setImage(defaultImagePath);
         memberRepository.save(member);
-        log.info("{} 유저 이미지 삭제 완료 - 기본 이미지로 변경 : {}", memberId, defaultImagePath);
+        log.info("{} 유저 이미지 삭제 완료 - 기본 이미지로 변경 : {}", id, defaultImagePath);
     }
 
     // 해당 아이디의 비밀번호를 새로운 비밀번호로 변경
@@ -216,12 +206,23 @@ public class MemberServiceImpl implements MemberService {
     }
 
     // 로그인 계정이 정보변경 계정과 일치하는지 체크
-    private void identifyLoginUser(String loginId, String id) {
+    public void checkAuthorizedUser(Principal principal, String id) {
+        String loginId = getLoginUserId(principal);
         if (!id.equals(loginId)) {
             log.info("로그인 계정과 정보변경 요청 계정 불일치 - login : {}, request : {}", loginId, id);
-            throw new CustomRuntimeException(HttpStatus.FORBIDDEN, ExceptionEnum.NOT_ALLOWED_ACCESS);
+            throw new CustomRuntimeException(HttpStatus.FORBIDDEN, ExceptionEnum.FORBIDDEN);
         }
         log.info("로그인 계정과 정보변경 요청 계정 일치 - login : {}, request : {}", loginId, id);
+    }
+
+    public String getLoginUserId(Principal principal) {
+        if (principal == null) {
+            log.info("로그인 상태 확인 불가");
+            throw new CustomRuntimeException(HttpStatus.UNAUTHORIZED, ExceptionEnum.UNAUTHORIZED);
+        }
+        String loginId = principal.getName();
+        log.info("로그인 상태 확인 - {} 유저", loginId);
+        return loginId;
     }
 
     // 기존 비밀번호가 변경될 비밀번호와 일치하지 않는지 체크
